Add initial index.html/GETTING-STARTED.html pages
The idea being that a user that has downloaded the prototype kit will be
able to get started by opening either page in their browser. The
index.html page will be the homepage for their prototype and they will
edit it. The GETTING-STARTED.html page will be the place they'll go to
to get help and learn how to work with the toolkit.